% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholupdate}
\alias{cholupdate}
\title{cholupdate}
\usage{
cholupdate(L, x, v, direction)
}
\arguments{
\item{L}{cholesky matrix which will be updated}

\item{x}{colvec which will be used for the update}

\item{v}{double: additional factor for cholupdate, see Van der Merwe and Wan, 2001}

\item{direction}{string: either downdate or update}
}
\value{
matrix: downdated / updated matrix
}
\description{
Computes the cholupdate for the square root unscented update (see Van der Merwe and Wan, 2001)
}
