% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clonePsydiffModel}
\alias{clonePsydiffModel}
\title{clonePsydiffModel}
\usage{
clonePsydiffModel(model)
}
\arguments{
\item{model}{psydiff model}
}
\value{
clone of the model
}
\description{
additional function to deep-copy a psydiffModel
}
