% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareModel.R
\name{compileModel}
\alias{compileModel}
\title{compileModel}
\usage{
compileModel(psydiffModel)
}
\arguments{
\item{psydiffModel}{model from newPsydiff}
}
\description{
compile the models from newPsydiff
}
