% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareModel.R
\name{compileParallelGradients}
\alias{compileParallelGradients}
\title{compileParallelGradients}
\usage{
compileParallelGradients(psydiffModel, nCores)
}
\arguments{
\item{psydiffModel}{model from newPsydiff}

\item{nCores}{number of cores to use}
}
\value{
psydiffModel with additional cl field for workers
}
\description{
compile the model from newPsydiff for use with parallel gradient computation. This function will invoke parallel::makeCluster(nCores, type = "PSOCK"). Stop the cluster with stopParallelGradients().
}
