% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeIndividualM2LL}
\alias{computeIndividualM2LL}
\title{computeIndividualM2LL}
\usage{
computeIndividualM2LL(
  nObservedVariables,
  rawData,
  expectedMeans,
  expectedCovariance
)
}
\arguments{
\item{nObservedVariables}{number of non-NA observations for this person}

\item{rawData}{colvec with observed data for this person}

\item{expectedMeans}{colvec with predicted manifest means for this person}

\item{expectedCovariance}{matrix with predicted manifest covariances for this person}
}
\value{
double: -2 log likelihood
}
\description{
Computes the -2 log likelihood for a single subject
}
