% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeK}
\alias{computeK}
\title{computeK}
\usage{
computeK(C, S)
}
\arguments{
\item{C}{matrix with predicted cross-covariances}

\item{S}{matrix with predicted manifest covariances}
}
\value{
matrix: Kalman Gain
}
\description{
Computes the Kalman Gain
}
