% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeMeanFromSigmaPoints}
\alias{computeMeanFromSigmaPoints}
\title{computeMeanFromSigmaPoints}
\usage{
computeMeanFromSigmaPoints(sigmaPoints, meanWeights)
}
\arguments{
\item{sigmaPoints}{sigma points}

\item{meanWeights}{matrix with mean weights from getMeanWeights}
}
\value{
colvec with means
}
\description{
Computes the means given the sigma points and the mean weights
}
