% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computePFromSigmaPoints}
\alias{computePFromSigmaPoints}
\title{computePFromSigmaPoints}
\usage{
computePFromSigmaPoints(sigmaPoints, meanWeights, c)
}
\arguments{
\item{sigmaPoints}{sigma points}

\item{meanWeights}{matrix with mean weights from getMeanWeights}

\item{c}{double: hyperparameter c for computing the sigma points (controls the spread)}
}
\value{
latent covariance matrix
}
\description{
Computes the latent covariance matrix given the sigma points and the covariance weights
}
