% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{extractOptimx}
\alias{extractOptimx}
\title{extractOptimx}
\usage{
extractOptimx(model, opt)
}
\arguments{
\item{model}{psydiff model}

\item{opt}{result from calling psydiffOptimx}
}
\description{
sets the model parameters to the best values obtained from optimx
}
