% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAFromSigmaPoints}
\alias{getAFromSigmaPoints}
\title{getAFromSigmaPoints}
\usage{
getAFromSigmaPoints(sigmaPoints, meanWeights, c)
}
\arguments{
\item{sigmaPoints}{sigma points}

\item{meanWeights}{matrix with mean weights from getMeanWeights}

\item{c}{double: hyperparameter c for computing the sigma points (controls the spread)}
}
\value{
lower triangular matrix
}
\description{
Computes the lower triangular covariance matrix given the sigma points and the covariance weights
}
