% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCovWeights}
\alias{getCovWeights}
\title{getCovWeights}
\usage{
getCovWeights(n, alpha, beta, kappa)
}
\arguments{
\item{n}{number of latent variables}

\item{alpha}{hyperparameter}

\item{beta}{hyperparameter}

\item{kappa}{hyperparameter}
}
\value{
matrix with covariance weights
}
\description{
Computes the covariance weights of the unscented Kalman filter
}
