% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{getDataTemplate}
\alias{getDataTemplate}
\title{getDataTemplate}
\usage{
getDataTemplate(nSubjects, nVariables, nTimpoints, dts)
}
\arguments{
\item{nSubjects}{number of persons}

\item{nVariables}{number of simulatedObservation variables}

\item{nTimpoints}{number of time points per person}

\item{dts}{either a single dt, a vector with dts for an individual or a vector with dts for all individuals}
}
\description{
Generate a template for data simulation
}
