% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareModel.R
\name{getGradientsParallel}
\alias{getGradientsParallel}
\title{getGradientsParallel}
\usage{
getGradientsParallel(psydiffModel)
}
\arguments{
\item{psydiffModel}{model from newPsydiff}
}
\description{
Computes gradients in parallel. Requires compileParallelGradients to be used first!
}
