% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMeanWeights}
\alias{getMeanWeights}
\title{getMeanWeights}
\usage{
getMeanWeights(n, alpha, beta, kappa)
}
\arguments{
\item{n}{number of latent variables}

\item{alpha}{hyperparameter}

\item{beta}{hyperparameter}

\item{kappa}{hyperparameter}
}
\value{
matrix with mean weights
}
\description{
Computes the mean weights of the unscented Kalman filter
}
