% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getParameterValues}
\alias{getParameterValues}
\title{Get the parameter values of a panelSDEModel}
\usage{
getParameterValues(panelSDEModel)
}
\arguments{
\item{panelSDEModel}{panelSDEModel}
}
\value{
named vector with parameter values
}
\description{
Get the parameter values of a panelSDEModel
}
