% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPhi}
\alias{getPhi}
\title{getPhi}
\usage{
getPhi(squareMatrix)
}
\arguments{
\item{squareMatrix}{a square matrix}
}
\value{
lower triangular matrix
}
\description{
Computes the lower triangular matrix in Formula (33) of Saerkkae (2007)
}
