% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIST.R
\name{getRegValue}
\alias{getRegValue}
\title{getRegValue}
\usage{
getRegValue(lambda, theta, regIndicators, adaptiveLassoWeights = NULL)
}
\arguments{
\item{lambda}{Penaltiy value}

\item{regIndicators}{Names of regularized parameters}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}
}
\description{
computes sum(lambda*abs(regularized Values))
}
