% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSigmaPoints}
\alias{getSigmaPoints}
\title{getSigmaPoints}
\usage{
getSigmaPoints(m, A, covarianceIsRoot, c)
}
\arguments{
\item{m}{colvec of latent means}

\item{A}{lower triangular Cholesky decomposition of latent covariance matrix}

\item{covarianceIsRoot}{boolean: is A the lower triangular Cholesky decomposition?}

\item{c}{double: hyperparameter c for computing the sigma points (controls the spread)}
}
\value{
matrix with sigma points
}
\description{
Computes the sigma-points of the unscented Kalman filter
}
