% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{getStandardErrors}
\alias{getStandardErrors}
\title{getStandardErrors}
\usage{
getStandardErrors(model, ...)
}
\arguments{
\item{model}{psydiff model}

\item{...}{additional arguments to pass to optimHess}
}
\value{
vector with standard errors
}
\description{
compute standard errors for a fitted model using the diagonal elements in the inverse of the hessian approximation from optimHess
}
