% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{getStandardErrorsHW}
\alias{getStandardErrorsHW}
\title{getStandardErrorsHW}
\usage{
getStandardErrorsHW(model, ...)
}
\arguments{
\item{model}{psydiff model}

\item{...}{additional arguments to pass to optimHess}
}
\value{
vector with standard errors
}
\description{
compute Huber White robust standard errors for a fitted model
}
