% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIST.R
\name{getSubgradients}
\alias{getSubgradients}
\title{getSubgradients}
\usage{
getSubgradients(
  theta,
  jacobian,
  regIndicators,
  lambda,
  lineSearch,
  adaptiveLassoWeightsMatrix
)
}
\arguments{
\item{theta}{vector with named parameters}

\item{jacobian}{derivative of L(theta)}

\item{regIndicators}{names of regularized parameters}

\item{lambda}{lambda value}
}
\description{
Computes the subgradients of a lasso penalized likelihood f(theta) = L(theta)+lambda*p(theta)
}
\details{
NOTE: Function located in file exact_optimization.R
}
