% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getWMatrix}
\alias{getWMatrix}
\title{Computes the weight matrix}
\usage{
getWMatrix(meanWeights, covWeights)
}
\arguments{
\item{meanWeights}{matrix with mean weights from getMeanWeights}

\item{covWeights}{matrix with covariance weights from getCovWeights}
}
\value{
matrix with weights
}
\description{
Computes the weight matrix
}
