% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logChol2Chol}
\alias{logChol2Chol}
\title{logChol2Chol}
\usage{
logChol2Chol(logChol)
}
\arguments{
\item{logChol}{log-Cholesky decomposition of a matrix}
}
\value{
Cholesky decomposition of the matrix
}
\description{
transforms the log-Cholesky decomposition of a matrix to the Cholesky (see Pinheiro, J. C., and Bates, D. M. (1996). Unconstrained parametrizations for variance-covariance matrices. Statistics and Computing, 6(3), 289–296)
}
