% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareModel.R
\name{panelsde}
\alias{panelsde}
\title{panelsde}
\usage{
panelsde(
  dataset,
  latentEquations,
  manifestEquations,
  L,
  Rchol,
  A0,
  m0,
  grouping = NULL,
  parameters,
  groupingvariables = NULL,
  additional = NULL,
  compile = TRUE
)
}
\description{
panelsde
}
