% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictC}
\alias{predictC}
\title{predictC}
\usage{
predictC(sigmaPoints, Y_, W)
}
\arguments{
\item{sigmaPoints}{sigma points}

\item{Y_}{predicted manifest values based on the sigma points}

\item{W}{matrix with covariance and mean weights computed with getWMatrix}
}
\value{
matrix with predicted cross-covariances
}
\description{
Computes the predicted cross-covariance between latent and manifest variables
}
