% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictMu}
\alias{predictMu}
\title{predictMu}
\usage{
predictMu(Y_, meanWeights)
}
\arguments{
\item{Y_}{predicted manifest values based on the sigma points}

\item{meanWeights}{matrix with mean weights from getMeanWeights}
}
\value{
colvec with predicted manifest means
}
\description{
Computes the predicted manifest means
}
