% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictS}
\alias{predictS}
\title{predictS}
\usage{
predictS(Y_, W, R)
}
\arguments{
\item{Y_}{predicted manifest values based on the sigma points}

\item{W}{matrix with covariance and mean weights computed with getWMatrix}

\item{R}{manifest covariance matrix}
}
\value{
matrix with predicted manifest covariances
}
\description{
Computes the predicted manifest covariance
}
