\name{psydiff-package}
\alias{psydiff-package}
\alias{psydiff}
\docType{package}
\title{\packageTitle{psydiff}}
\description{\packageDescription{psydiff}}
\section{Package Content}{\packageIndices{psydiff}}
\author{\packageAuthor{psydiff}}
\section{Maintainer}{\packageMaintainer{psydiff}}
\keyword{package}
