% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffDEoptim}
\alias{psydiffDEoptim}
\title{psydiffDEoptim}
\usage{
psydiffDEoptim(model, lower, upper, ...)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffDEoptim)}

\item{lower}{vector specifying a lower bound for each parameter in the model}

\item{upper}{vector specifying an upper bound for each parameter in the model}

\item{...}{additional parameters to pass to DEoptim}
}
\description{
wrapper for DEoptim optimization
}
