% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffFitInternal}
\alias{psydiffFitInternal}
\title{psydiffFitInternal}
\usage{
psydiffFitInternal(pars, parsnames, model)
}
\arguments{
\item{pars}{vector with parameter values}

\item{parsnames}{vector with parameter names}

\item{model}{psydiff model object (must have been compiled before calling psydiffDEoptim)}
}
\description{
internal function for optimization
}
