% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffGradientsInternal}
\alias{psydiffGradientsInternal}
\title{psydiffGradientsInternal}
\usage{
psydiffGradientsInternal(pars, parsnames, model)
}
\arguments{
\item{pars}{vector with parameter values}

\item{parsnames}{vector with parameter names}

\item{model}{psydiff model object}
}
\description{
internal function for optimization
}
