% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffNLopt}
\alias{psydiffNLopt}
\title{psydiffNLopt}
\usage{
psydiffNLopt(model, eval_f = psydiffFitInternal, ...)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffNLopt)}

\item{...}{additional parameters to pass to NLopt}
}
\description{
wrapper for NLopt optimization
}
