% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffOptimBFGS}
\alias{psydiffOptimBFGS}
\title{psydiffOptimBFGS}
\usage{
psydiffOptimBFGS(model, ...)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffOptimBFGS)}

\item{...}{additional parameters to pass to optim}
}
\description{
wrapper for BFGS optimizer from optim
}
