% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffOptimNelderMead}
\alias{psydiffOptimNelderMead}
\title{psydiffOptimNelderMead}
\usage{
psydiffOptimNelderMead(model, ...)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffOptimNelderMead)}

\item{...}{additional parameters to pass to optim}
}
\description{
wrapper for Nelder-Mead optimizer from optim
}
