% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffOptimx}
\alias{psydiffOptimx}
\title{psydiffOptimx}
\usage{
psydiffOptimx(model, method = c("Nelder-Mead", "BFGS", "nlm", "nlminb"), ...)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffOptimx)}

\item{...}{additional parameters to pass to optimx}
}
\description{
wrapper for optimx optimization
}
