% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffOptimxMultiStart}
\alias{psydiffOptimxMultiStart}
\title{psydiffOptimxMultiStart}
\usage{
psydiffOptimxMultiStart(
  model,
  method = c("Nelder-Mead", "BFGS"),
  startMatrix = NULL,
  nstart = 20,
  sd = 1,
  ...
)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffOptimx)}

\item{...}{additional parameters to pass to optimx}
}
\description{
wrapper for optimx optimization
}
