% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimWrapper.R
\name{psydiffSubplex}
\alias{psydiffSubplex}
\title{psydiffSubplex}
\usage{
psydiffSubplex(model, ...)
}
\arguments{
\item{model}{psydiff model object (must have been compiled before calling psydiffSubplex)}

\item{...}{additional parameters to pass to subplex}
}
\description{
wrapper for subplex optimization
}
