% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr_}
\alias{qr_}
\title{qr_}
\usage{
qr_(X)
}
\arguments{
\item{X}{matrix which will be decomposed}
}
\value{
matrix: R-tilde
}
\description{
qr for the square root unscented update (see van der Merwe & Wan, 2001). Returns R-tilde
}
