% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setParameterList}
\alias{setParameterList}
\title{The parameterList holds all parameters for a single person.
With setParameterTable these parameters are changed to the
values specified in the parameterTable for a person}
\usage{
setParameterList(parameterTable, parameterList, person)
}
\arguments{
\item{parameterTable}{parameterTable}

\item{parameterList}{list of parameters}

\item{person}{integer indicating the person for which the parameters should be written in the parameterList}
}
\value{
the parameterList is changed by reference; nothing is returned explicitly
}
\description{
The parameterList holds all parameters for a single person.
With setParameterTable these parameters are changed to the
values specified in the parameterTable for a person
}
