% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setParameterValues}
\alias{setParameterValues}
\title{Change the parameter values in the parameterTable}
\usage{
setParameterValues(parameterTable, parameterValues, parameterLabels)
}
\arguments{
\item{parameterTable}{parameterTable}

\item{parameterValues}{values to which the parameters in parameterTable will be changed}

\item{parameterLabels}{labels of the parameters (must correspond to the label column of parameterTable)}
}
\value{
the parameterTable is changed and returned
}
\description{
Change the parameter values in the parameterTable
}
