% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareModel.R
\name{stopParallelGradients}
\alias{stopParallelGradients}
\title{stopParallelGradients}
\usage{
stopParallelGradients(psydiffModel)
}
\arguments{
\item{psydiffModel}{model from newPsydiff}
}
\value{

}
\description{
Stops the workers.
}
