% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateM}
\alias{updateM}
\title{updateM}
\usage{
updateM(m_, K, residual)
}
\arguments{
\item{m_}{colvec with predicted latent means}

\item{K}{Kalman Gain matrix}

\item{residual}{colvec with difference between predicted manifest means (mu) and observed manifest values}
}
\value{
colvec with updated latent means
}
\description{
Updates the latent means
}
