% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateP}
\alias{updateP}
\title{updateP}
\usage{
updateP(P_, K, S)
}
\arguments{
\item{P_}{matrix with predicted latent covariances}

\item{K}{Kalman Gain matrix}

\item{S}{matrix with predicted manifest covariances}
}
\value{
matrix with updated latent covariances
}
\description{
Updates the latent covariances
}
