% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getParameterValues}
\alias{getParameterValues}
\title{getParameterValues}
\usage{
getParameterValues(psydiffModel)
}
\arguments{
\item{psydiffModel}{psydiffModel}
}
\value{
named vector with parameter values
}
\description{
Get the parameter values of a psydiffModel
}
